New Features
Added the ability to revert changes one at a time instead of all at once, and allowed users to view a list before it being added to the catalog of their own. A few issues though, and that's that only a single task in a list can display the revert option at a time, and that the button to add a list to your own is out of line with the main body due to how I nested the elements originally.